+++ os16/install.sh
@@ -126,12 +126,13 @@ write_os16_ai_prop() {
   blurv2=0
   exp=0
   launcher_async=0
+  # Blur tiers: 1=Smart solid, 2=partial (dock blur, solid shade), 3=full glass.
   if [ "$blur" = "true" ]; then
     case "$blvl" in
       1) b01=0; sfdis=1; gblur=0; dynblur=0; blurv2=0; exp=0; launcher_async=0; union=0 ;;
-      2) b01=1; sfdis=0; gblur=2; dynblur=0; blurv2=0; exp=0; launcher_async=$a01; union=$a01 ;;
+      2) b01=0; sfdis=1; gblur=2; dynblur=0; blurv2=0; exp=0; launcher_async=$a01; union=0 ;;
       3) b01=1; sfdis=0; gblur=3; dynblur=1; blurv2=1; exp=1; launcher_async=$a01; union=$a01 ;;
-      *) b01=1; sfdis=0; gblur=2; dynblur=0; blurv2=0; exp=0; launcher_async=$a01; union=$a01 ;;
+      *) b01=0; sfdis=1; gblur=2; dynblur=0; blurv2=0; exp=0; launcher_async=$a01; union=0 ;;
     esac
   else
     b01=0; sfdis=1; gblur=0; dynblur=0; blurv2=0; exp=0; launcher_async=0; union=0